requirements.txt
--only-binary pymupdf

cached_property
click
defusedxml>=0.6.0
html2text
https://github.com/turicas/rows/archive/develop.zip#egg=rows
jinja2
openpyxl
python-Levenshtein
pytz
requests
s3cmd
scrapy
tqdm
goodtables==2.4.9

gunicorn
flask
epiweeks
